fixed graph type in sparse matrix

+++ linalg/sparsemat.cpp
@@ -186,16 +186,16 @@ SparseMatrix::SparseMatrix(const SparseMatrix &mat, bool copy_graph)
 void SparseMatrix::Push() const
 {
   const int nnz = I[height];
-  mm::push(I, (height+1)*sizeof(double));
-  mm::push(J, nnz*sizeof(double));
+  mm::push(I, (height+1)*sizeof(int));
+  mm::push(J, nnz*sizeof(int));
   mm::push(A, nnz*sizeof(double));
 }
 
 void SparseMatrix::Pull() const
 {
   const int nnz = I[height];
-  mm::pull(I, (height+1)*sizeof(double));
-  mm::pull(J, nnz*sizeof(double));
+  mm::pull(I, (height+1)*sizeof(int));
+  mm::pull(J, nnz*sizeof(int));
   mm::pull(A, nnz*sizeof(double));
 }
